minor refactoring & little bit more modularity and ability to add more routers
