+++ src/main/scala/com/campudus/tableaux/database/DatabaseAccess.scala
@@ -1,0 +1,8 @@
+package com.campudus.tableaux.database
+
+/**
+ * Convenient trait for DatabaseAccess
+ */
+trait DatabaseAccess {
+  val database: DatabaseConnection
+}
