+++ src/main/scala/com/campudus/tableaux/database/DatabaseQuery.scala
@@ -1,12 +1,17 @@
 package com.campudus.tableaux.database
 
 import scala.concurrent.Future
-import org.vertx.scala.core.json.{ JsonObject, JsonArray, Json }
+import org.vertx.scala.core.json.{ JsonObject, Json }
 import com.campudus.tableaux.database.Tableaux._
 import com.campudus.tableaux.database.ResultChecker._
 
-class SystemStructure(connection: DatabaseConnection) {
+trait DatabaseQuery {
+  protected[this] val connection: DatabaseConnection
+
   implicit val executionContext = connection.executionContext
+}
+
+class SystemStructure(protected val connection: DatabaseConnection) extends DatabaseQuery {
 
   def deinstall(): Future[Unit] = for {
     t <- connection.begin()
@@ -66,13 +71,12 @@ class SystemStructure(connection: DatabaseConnection) {
   } yield ()
 }
 
-class TableStructure(connection: DatabaseConnection) {
-  implicit val executionContext = connection.executionContext
+class TableStructure(val connection: DatabaseConnection) extends DatabaseQuery {
 
   def create(name: String): Future[IdType] = for {
     t <- connection.begin()
     (t, result) <- t.query("INSERT INTO system_table (user_table_name) VALUES (?) RETURNING table_id", Json.arr(name))
-    id <- Future.apply(insertNotNull(result)(0).get[Long](0))
+    id <- Future.apply(insertNotNull(result).head.get[Long](0))
     (t, _) <- t.query(s"CREATE TABLE user_table_$id (id BIGSERIAL, PRIMARY KEY (id))", Json.arr())
     (t, _) <- t.query(s"CREATE SEQUENCE system_columns_column_id_table_$id", Json.arr())
     _ <- t.commit()
@@ -85,7 +89,7 @@ class TableStructure(connection: DatabaseConnection) {
   def get(tableId: IdType): Future[(IdType, String)] = {
     connection.singleQuery("SELECT table_id, user_table_name FROM system_table WHERE table_id = ?", Json.arr(tableId))
   } map { r =>
-    val json = selectNotNull(r)(0)
+    val json = selectNotNull(r).head
     (json.get[IdType](0), json.get[String](1))
   }
 
@@ -103,8 +107,7 @@ class TableStructure(connection: DatabaseConnection) {
   } map (_ => ())
 }
 
-class ColumnStructure(connection: DatabaseConnection) {
-  implicit val executionContext = connection.executionContext
+class ColumnStructure(val connection: DatabaseConnection) extends DatabaseQuery {
 
   def insert(tableId: IdType, dbType: TableauxDbType, name: String, ordering: Option[Ordering]): Future[(IdType, Ordering)] = for {
     t <- connection.begin()
@@ -120,7 +123,7 @@ class ColumnStructure(connection: DatabaseConnection) {
                      |  RETURNING column_id, ordering""".stripMargin,
         Json.arr(tableId, dbType.toString, name, ord))
     }
-    result <- Future.successful(insertNotNull(result)(0))
+    result <- Future.successful(insertNotNull(result).head)
     (t, _) <- t.query(s"ALTER TABLE user_table_$tableId ADD column_${result.get[IdType](0)} $dbType", Json.arr())
     _ <- t.commit()
   } yield (result.get[IdType](0), result.get[Ordering](1))
@@ -129,7 +132,7 @@ class ColumnStructure(connection: DatabaseConnection) {
     t <- connection.begin()
     (t, result) <- t.query(s"""INSERT INTO system_link_table (table_id_1, table_id_2, column_id_1, column_id_2) VALUES (?, ?, ?, ?) RETURNING link_id""".stripMargin,
       Json.arr(tableId, toTableId, fromColumnId, toColumnId))
-    linkId <- Future.successful { insertNotNull(result)(0).get[IdType](0) }
+    linkId <- Future.successful { insertNotNull(result).head.get[IdType](0) }
     (t, _) <- t.query(s"""
                     |INSERT INTO system_columns (table_id, column_id, column_type, user_column_name, ordering, link_id) VALUES (
                     |  ?, 
@@ -177,7 +180,7 @@ class ColumnStructure(connection: DatabaseConnection) {
                     |)""".stripMargin, Json.arr())
     _ <- t.commit()
   } yield {
-    val json = insertNotNull(result)(0)
+    val json = insertNotNull(result).head
     (json.get[IdType](0), json.get[Ordering](1))
   }
 
@@ -188,7 +191,7 @@ class ColumnStructure(connection: DatabaseConnection) {
                               |  WHERE table_id = ? AND column_id = ?
                               |  ORDER BY column_id""".stripMargin, Json.arr(tableId, columnId))
   } yield {
-    val json = selectNotNull(result)(0)
+    val json = selectNotNull(result).head
     (json.get[IdType](0), json.get[String](1), Mapper.getDatabaseType(json.get[String](2)), json.get[Ordering](3))
   }
 
@@ -203,7 +206,7 @@ class ColumnStructure(connection: DatabaseConnection) {
                               |  WHERE table_id != ? AND column_id != ? 
                               |  ORDER BY column_id""".stripMargin, Json.arr(tableId, columnId))
   } yield {
-    val json = selectNotNull(result)(0)
+    val json = selectNotNull(result).head
     (json.get[IdType](0), json.get[IdType](1))
   }
 
@@ -219,7 +222,7 @@ class ColumnStructure(connection: DatabaseConnection) {
     t <- connection.begin()
     (t, result1) <- optionToValidFuture(columnName, t, { name: String => t.query(s"UPDATE system_columns SET user_column_name = ? WHERE table_id = ? AND column_id = ?", Json.arr(name, tableId, columnId)) })
     (t, result2) <- optionToValidFuture(ordering, t, { ord: Ordering => t.query(s"UPDATE system_columns SET ordering = ? WHERE table_id = ? AND column_id = ?", Json.arr(ord, tableId, columnId)) })
-    (t, result3) <- optionToValidFuture(kind, t, { k: TableauxDbType => t.query(s"UPDATE system_columns SET column_type = ? WHERE table_id = ? AND column_id = ?", Json.arr(k.toString(), tableId, columnId)) })
+    (t, result3) <- optionToValidFuture(kind, t, { k: TableauxDbType => t.query(s"UPDATE system_columns SET column_type = ? WHERE table_id = ? AND column_id = ?", Json.arr(k.toString, tableId, columnId)) })
     (t, _) <- optionToValidFuture(kind, t, { k: TableauxDbType => t.query(s"ALTER TABLE user_table_$tableId ALTER COLUMN column_$columnId TYPE ${k.toString} USING column_$columnId::${k.toString}", Json.arr()) })
     _ <- Future.apply(checkUpdateResults(Seq(result1, result2, result3))) recoverWith t.rollbackAndFail()
     _ <- t.commit()
@@ -235,32 +238,31 @@ class ColumnStructure(connection: DatabaseConnection) {
   }
 }
 
-class RowStructure(connection: DatabaseConnection) {
-  implicit val executionContext = connection.executionContext
+class RowStructure(val connection: DatabaseConnection) extends DatabaseQuery {
 
   def create(tableId: IdType): Future[IdType] = {
     connection.singleQuery(s"INSERT INTO user_table_$tableId DEFAULT VALUES RETURNING id", Json.arr())
-  } map { insertNotNull(_)(0).get[IdType](0) }
+  } map { insertNotNull(_).head.get[IdType](0) }
 
   def createFull(tableId: IdType, values: Seq[(IdType, _)]): Future[IdType] = {
     val qm = values.foldLeft(Seq[String]())((s, _) => s :+ "?").mkString(", ")
     val columns = values.foldLeft(Seq[String]())((s, tup) => s :+ s"column_${tup._1}").mkString(", ")
     val v = values.foldLeft(Seq[Any]())((s, tup) => s :+ tup._2)
     connection.singleQuery(s"INSERT INTO user_table_$tableId ($columns) VALUES ($qm) RETURNING id", Json.arr(v: _*))
-  } map { insertNotNull(_)(0).get[IdType](0) }
+  } map { insertNotNull(_).head.get[IdType](0) }
 
   def get(tableId: IdType, rowId: IdType): Future[(IdType, Seq[AnyRef])] = {
     connection.singleQuery(s"SELECT * FROM user_table_$tableId WHERE id = ?", Json.arr(rowId))
   } map { x =>
-    val seq = jsonArrayToSeq(selectNotNull(x)(0))
-    (seq(0), seq.drop(1))
+    val seq = jsonArrayToSeq(selectNotNull(x).head)
+    (seq.head, seq.drop(1))
   }
 
   def getAll(tableId: IdType): Future[Seq[(IdType, Seq[AnyRef])]] = {
     connection.singleQuery(s"SELECT * FROM user_table_$tableId ORDER BY id", Json.arr())
   } map { x =>
-    val seq = getSeqOfJsonArray(x) map { jsonArrayToSeq(_) }
-    seq map { s => (s(0), s.drop(1)) }
+    val seq = getSeqOfJsonArray(x) map { jsonArrayToSeq }
+    seq map { s => (s.head, s.drop(1)) }
   }
 
   def delete(tableId: IdType, rowId: IdType): Future[Unit] = {
@@ -268,8 +270,7 @@ class RowStructure(connection: DatabaseConnection) {
   } map { deleteNotNull(_) }
 }
 
-class CellStructure(connection: DatabaseConnection) {
-  implicit val executionContext = connection.executionContext
+class CellStructure(val connection: DatabaseConnection) extends DatabaseQuery {
 
   def update[A](tableId: IdType, columnId: IdType, rowId: IdType, value: A): Future[Unit] = {
     connection.singleQuery(s"UPDATE user_table_$tableId SET column_$columnId = ? WHERE id = ?", Json.arr(value, rowId))
@@ -278,22 +279,22 @@ class CellStructure(connection: DatabaseConnection) {
   def updateLink(tableId: IdType, linkColumnId: IdType, values: (IdType, IdType)): Future[Unit] = for {
     t <- connection.begin()
     (t, result) <- t.query("SELECT link_id FROM system_columns WHERE table_id = ? AND column_id = ?", Json.arr(tableId, linkColumnId))
-    linkId <- Future.successful(selectNotNull(result)(0).get[IdType](0))
+    linkId <- Future.successful(selectNotNull(result).head.get[IdType](0))
     (t, _) <- t.query(s"INSERT INTO link_table_$linkId VALUES (?, ?)", Json.arr(values._1, values._2))
     _ <- t.commit()
   } yield ()
 
   def getValue(tableId: IdType, columnId: IdType, rowId: IdType): Future[Any] = {
     connection.singleQuery(s"SELECT column_$columnId FROM user_table_$tableId WHERE id = ?", Json.arr(rowId))
-  } map { selectNotNull(_)(0).get[Any](0) }
+  } map { selectNotNull(_).head.get[Any](0) }
 
   def getLinkValues(tableId: IdType, linkColumnId: IdType, rowId: IdType, toTableId: IdType, toColumnId: IdType): Future[Any] = for {
     t <- connection.begin()
     (t, result) <- t.query("SELECT link_id FROM system_columns WHERE table_id = ? AND column_id = ?", Json.arr(tableId, linkColumnId))
-    linkId <- Future.successful(selectNotNull(result)(0).get[IdType](0))
+    linkId <- Future.successful(selectNotNull(result).head.get[IdType](0))
     (t, result) <- t.query("SELECT table_id_1, table_id_2, column_id_1, column_id_2 FROM system_link_table WHERE link_id = ?", Json.arr(linkId))
     (id1, id2) <- Future.successful {
-      val res = selectNotNull(result)(0)
+      val res = selectNotNull(result).head
       val linkTo2 = (res.get[IdType](1), res.get[IdType](3))
 
       if (linkTo2 == (toTableId, toColumnId)) ("id_1", "id_2") else ("id_2", "id_1")
@@ -301,11 +302,11 @@ class CellStructure(connection: DatabaseConnection) {
     (t, result) <- t.query(s"""
                      |SELECT STRING_AGG(user_table_$toTableId.column_$toColumnId, ', ') FROM user_table_$tableId 
                      |  LEFT JOIN link_table_$linkId 
-                     |    ON user_table_$tableId.id = link_table_${linkId}.$id1
+                     |    ON user_table_$tableId.id = link_table_$linkId.$id1
                      |  LEFT JOIN user_table_$toTableId 
-                     |    ON user_table_$toTableId.id = link_table_${linkId}.$id2
+                     |    ON user_table_$toTableId.id = link_table_$linkId.$id2
                      |  WHERE user_table_$tableId.id = ?
                      |  GROUP BY user_table_$tableId.id""".stripMargin, Json.arr(rowId))
     _ <- t.commit()
-  } yield selectNotNull(result)(0).get[Any](0)
+  } yield selectNotNull(result).head.get[Any](0)
 }
