+++ src/main/scala/com/campudus/tableaux/database/ResultChecker.scala
@@ -24,6 +24,10 @@ object ResultChecker {
   def updateNotNull(json: JsonObject): Seq[JsonArray] = checkHelper(json, "UPDATE 0", "update")
 
   private def checkHelper(json: JsonObject, s: String, ex: String): Seq[JsonArray] = {
-    if (json.getString("message") == s) throw NotFoundInDatabaseException(s"Warning: $ex query failed", ex) else getSeqOfJsonArray(json)
+    if (json.getString("message") == s) {
+      throw NotFoundInDatabaseException(s"Warning: $ex query failed", ex)
+    } else {
+      getSeqOfJsonArray(json)
+    }
   }
 }
