+++ src/test/scala/com/campudus/tableaux/TableauxTestBase.scala
@@ -1,16 +1,17 @@
 package com.campudus.tableaux
 
+
 import org.vertx.scala.core.FunctionConverters._
-import org.vertx.scala.core.json.{ Json, JsonObject }
+import org.vertx.scala.core.json._
 import org.vertx.scala.testtools.TestVerticle
 import org.vertx.testtools.VertxAssert._
 import org.vertx.scala.core.http._
 import scala.concurrent.{ Promise, Future }
+import scala.io.Source
+import java.io.File
 import scala.util.{ Try, Failure, Success }
 import com.campudus.tableaux.database.SystemStructure
 import com.campudus.tableaux.database.DatabaseConnection
-import org.vertx.scala.core.buffer.Buffer
-import org.vertx.scala.router.RouterException
 
 case class TestCustomException(message: String, id: String, statusCode: Int) extends Throwable
 
@@ -20,19 +21,25 @@ case class TestCustomException(message: String, id: String, statusCode: Int) ext
 trait TableauxTestBase extends TestVerticle {
   val DEFAULT_PORT = 8181
 
+  lazy val config: JsonObject = jsonFromFile("../conf.json")
+  lazy val databaseAddress: String = config.getObject("database", Json.obj()).getString("address", "campudus.asyncdb")
+
+  private def readJsonFile(f: String): String = Source.fromFile(s"$f").getLines().mkString
+  private def jsonFromFile(f: String): JsonObject = Json.fromObjectString(readJsonFile(f))
+  
   override def asyncBefore(): Future[Unit] = {
-    val transaction = new DatabaseConnection(this)
+    val transaction = new DatabaseConnection(this, databaseAddress)
     val system = new SystemStructure(transaction)
     for {
-      _ <- deployModule()
+      _ <- deployModule(config)
       _ <- system.deinstall()
       _ <- system.setup()
     } yield ()
   }
 
-  def deployModule(): Future[Unit] = {
+  def deployModule(config: JsonObject): Future[Unit] = {
     val p = Promise[Unit]()
-    container.deployModule(System.getProperty("vertx.modulename"), Json.obj(), 1, {
+    container.deployModule(System.getProperty("vertx.modulename"), config, 1, {
       case Success(id) => p.success()
       case Failure(ex) =>
         logger.error("could not deploy", ex)
@@ -66,16 +73,18 @@ trait TableauxTestBase extends TestVerticle {
     }
   }
 
-  def createClient(): HttpClient = vertx.createHttpClient().setHost("localhost").setPort(DEFAULT_PORT)
+  def createClient(): HttpClient = {
+    vertx.createHttpClient().setHost("localhost").setPort(config.getInteger("port", DEFAULT_PORT))
+  }
 
   def sendRequest(method: String, path: String): Future[JsonObject] = {
-    val p = Promise[JsonObject]
+    val p = Promise[JsonObject]()
     httpClientRequest(method, path, p).end()
     p.future
   }
 
   def sendRequestWithJson(method: String, jsonObj: JsonObject, path: String): Future[JsonObject] = {
-    val p = Promise[JsonObject]
+    val p = Promise[JsonObject]()
     httpClientRequest(method, path, p).setChunked(true).write(jsonObj.encode()).end()
     p.future
   }
@@ -89,7 +98,7 @@ trait TableauxTestBase extends TestVerticle {
         p.failure(TestCustomException(buf.toString(), resp.statusMessage(), resp.statusCode()))
       } else {
         try {
-          p.success(Json.fromObjectString(buf.toString))
+          p.success(Json.fromObjectString(buf.toString()))
         } catch {
           case ex: Exception => p.failure(ex)
         }
